- Initial attempt to move Token HUD side menu rendering outside of token-variants.mjs
